// Meteor.js/8. Roles And Logging/imports/api/utils/models.ts
import { MethodSearchModel } from '/imports/types/interfaces';

export enum AvailableCollectionNames {
    POSTS = 'posts',
    POST_LIKES = 'post_likes',
    SYSTEM_CHANGE_LOGS = 'system_change_logs',
    USERS = 'users',
    USER_PROFILE = 'user_profiles',
}

export interface MethodUtilMethodsFindCollectionModel extends MethodSearchModel {
    collection: AvailableCollectionNames;
    /**
     * If provided, deleted documents will also be provided. If `onlyOne` is used, below will not be used.
     */
    includeDeleted?: boolean;
    count?: boolean;
}